feat(common): make unified Credentials public
Expose the public APIs for `google::cloud::Credentials`. Currently there
are no public APIs to *accept* these credentials in any libraries, that
will come in separate PRs.
