+++ google/cloud/CMakeLists.txt
@@ -110,6 +110,8 @@ add_library(
     ${CMAKE_CURRENT_BINARY_DIR}/internal/build_info.cc
     backoff_policy.h
     common_options.h
+    credentials.cc
+    credentials.h
     future.h
     future_generic.h
     future_void.h
@@ -132,8 +134,8 @@ add_library(
     internal/build_info.h
     internal/compiler_info.cc
     internal/compiler_info.h
-    internal/credentials.cc
-    internal/credentials.h
+    internal/credentials_impl.cc
+    internal/credentials_impl.h
     internal/diagnostics_pop.inc
     internal/diagnostics_push.inc
     internal/disable_deprecation_warnings.inc
@@ -260,7 +262,7 @@ if (BUILD_TESTING)
         internal/backoff_policy_test.cc
         internal/big_endian_test.cc
         internal/compiler_info_test.cc
-        internal/credentials_test.cc
+        internal/credentials_impl_test.cc
         internal/env_test.cc
         internal/filesystem_test.cc
         internal/format_time_point_test.cc
