+++ google/cloud/google_cloud_cpp_common_unit_tests.bzl
@@ -28,7 +28,7 @@ google_cloud_cpp_common_unit_tests = [
     "internal/backoff_policy_test.cc",
     "internal/big_endian_test.cc",
     "internal/compiler_info_test.cc",
-    "internal/credentials_test.cc",
+    "internal/credentials_impl_test.cc",
     "internal/env_test.cc",
     "internal/filesystem_test.cc",
     "internal/format_time_point_test.cc",
